Add highlevel debug prints in MSI-X handler
The problem seems to appear only when MSI-X is loaded. If regular MSI also
is loaded does not seem to matter. Since the printouts seem fine after the
memcpy_s for MSI-X call in populate_msi_struct seems fine, the next thing
to check is the handler that is registred for MSI-X.