Set O1 and retract all optimization flags explicit
Turns out I was not sane the last time. So a new approach. Set -O1
then remove all optimization flags that O1 would include. This should
be the same thing as specifying -O0. If this boots then I will continue
with binarysearching for the case where it no longer boots by allowing
subsets of the optimization flags.